# radio-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radiology API",
        "description": "Appointment booking application for a radiology department",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://radiologie-rdv.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/2fa/enable": {
            "post": {
                "tags": [
                    "2FA"
                ],
                "summary": "Activer la validation à deux facteurs",
                "description": "Valide le code 6 chiffres envoyé par l'utilisateur et active 2FA.",
                "operationId": "07e594577c5c9d0dc9b95dcf10d8581c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection vers le dashboard avec succès"
                    },
                    "422": {
                        "description": "Code invalide ou erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/2fa/disable": {
            "post": {
                "tags": [
                    "2FA"
                ],
                "summary": "Désactiver la validation à deux facteurs",
                "description": "Désactive 2FA pour l'utilisateur authentifié.",
                "operationId": "13efff48daa375b5508e0c8eec35e1dd",
                "responses": {
                    "302": {
                        "description": "Redirection vers le dashboard avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/2fa/setup": {
            "get": {
                "tags": [
                    "2FA"
                ],
                "summary": "Afficher la page de configuration 2FA",
                "description": "Retourne la vue avec QR code et clé secrète pour configurer Google Authenticator.",
                "operationId": "67cbab6845dbce7b84b101789e210d23",
                "responses": {
                    "200": {
                        "description": "Page de configuration 2FA affichée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/user/enable2fa": {
            "get": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Afficher la page d'activation 2FA",
                "description": "Génère une clé secrète 2FA, crée un QR code inline et retourne la vue d'activation pour l'utilisateur connecté.",
                "operationId": "0f9ea3ca413b0e3c28e62339f4186867",
                "responses": {
                    "200": {
                        "description": "Page d'activation 2FA affichée avec QR code"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}